requirements.txt
deep_translator==1.11.4
Django==5.2.1
pymongo==4.13.0
python-dotenv==1.1.0
Requests==2.32.3
